+++ src/common/utils/signers.js
@@ -38,12 +38,20 @@ class EnhancedWallet extends Wallet {
       return super.getTransactionCount(...args);
     return this._options.getTransactionCount(...args);
   }
+
+  sendTransaction(tx) {
+    let gasPrice;
+    if (this._options.gasPrice !== undefined) {
+      gasPrice = BigNumber.from(this._options.gasPrice).toHexString();
+    }
+    return super.sendTransaction({ gasPrice, ...tx });
+  }
 }
 
 class EnhancedWeb3Signer extends Signer {
-  constructor(web3) {
+  constructor(...args) {
     super();
-    const web3Provider = new Web3Provider(web3);
+    const web3Provider = new Web3Provider(...args);
     this.provider = web3Provider;
   }