+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 All notable changes to this project will be documented in this file.
 
+## [7.0.0] 2021-12-17
+
+### Added
+
+- ENS resolution on iExec sidechains and custom networks
+- ENS methods in CLI & lib:
+  - `iexec ens resolve <name>` and `iexec.ens.resolveName(name)`
+  - `iexec ens lookup <address>` and `iexec.ens.lookupAddress(addres)`
+  - `iexec ens get-owner <name>` and `iexec.ens.getOwner(name)`
+  - `iexec ens register <label>`
+  - `iexec.ens.claimName(label, domain)`
+  - `iexec.ens.configureResolution(label, address)`
+
+### Changed
+
+- fix `iexec registry validate app` to support new mrenclave format
+
+### Removed
+
 ## [6.0.0] 2021-07-19
 
 ### Added